+++ packages/sdk/src/sdk-blockchains/ethereum/bid.test.ts
@@ -13,10 +13,10 @@ import { createEthereumSdk } from "./index"
 
 describe("transfer", () => {
 
-	const { web31 } = initProviders({})
+	const { web31, wallet1 } = initProviders({})
 
 	const senderEthereum = new Web3Ethereum({ web3: web31 })
-	const senderSdk = createEthereumSdk(new EthereumWallet(senderEthereum, "e2e"))
+	const senderSdk = createEthereumSdk(new EthereumWallet(senderEthereum, toUnionAddress(wallet1.getAddressString()), "e2e"))
 
 	const raribleSdk = createRaribleSdk(senderEthereum, "e2e")